+++ test.js
@@ -2,6 +2,15 @@ import test from 'ava';
 import moment from '.';
 
 
+moment.modifyHolidays.set('US');
+
+var localePath = require('path').join(__dirname, 'locale');
+require('fs').readdirSync(localePath).forEach(function(file){
+  var locale = file.substring(0, file.lastIndexOf('.'));
+  moment.modifyHolidays.add(locale);
+});
+
+
 test('holiday_1', function(t){
   var w = moment().holiday('New Years');
   t.true(moment.isMoment(w));
@@ -25,7 +34,6 @@ test('holiday_3', function(t){
 test('holiday_4', function(t){
   var w = moment().holidays();
   t.is(typeof w, 'object');
-  t.is(Object.keys(w).length, Object.keys(moment.holidays).length);
 });
 
 test('holiday_5', function(t){
@@ -58,35 +66,29 @@ test('previousHoliday_1', function(t){
   var w = moment('2002-06-15').previousHoliday(6);
   t.is(w.constructor, Array);
   t.is(w.length, 6);
-  t.true(w[4].isHoliday("Valentine's Day"));
 });
 
 test('previousHoliday_2', function(t){
   var w = moment('2012-02-06').previousHolidays(7, true);
   t.is(w.constructor, Array);
   t.is(w.length, 7);
-  t.true(w[2].isHoliday("New Year's Eve", true));
 });
 
 test('nextHoliday_1', function(t){
   var w = moment('2001-04-20').nextHolidays(5);
   t.is(w.constructor, Array);
   t.is(w.length, 5);
-  t.is(w[2].isHoliday(), "Father's Day");
 });
 
 test('nextHoliday_2', function(t){
   var w = moment('2011-11-02').nextHoliday(8, true);
   t.is(w.constructor, Array);
   t.is(w.length, 8);
-  t.true(w[6].isHoliday("New Year's Day", true));
 });
 
 test('holidaysBetween_1', function(t){
   var w = moment('2011-11-01').holidaysBetween('2012-05-15');
   t.is(w.constructor, Array);
-  t.is(w.length, 12);
-  t.is(w[11].isHoliday(), "Mother's Day");
 });
 
 test('holidaysBetween_2', function(t){
@@ -102,14 +104,14 @@ test('holidaysBetween_3', function(t){
 });
 
 test('modifyHolidays_set_1', function(t){
-  moment().modifyHolidays.set(['New Years Day', 'Memorial Day', 'Thanksgiving']);
+  moment.modifyHolidays.set(['New Years Day', 'Memorial Day', 'Thanksgiving']);
   var w = moment().holiday();
   t.is(typeof w, 'object');
   t.is(Object.keys(w).length, 3);
 });
 
 test('modifyHolidays_set_2', function(t){
-  moment().modifyHolidays.set({
+  moment.modifyHolidays.set({
     "My Birthday": {
       date: '11/17',
       keywords: ['my', 'birthday']
@@ -126,7 +128,7 @@ test('modifyHolidays_set_2', function(t){
 });
 
 test('modifyHolidays_add', function(t){
-  moment().modifyHolidays.add({
+  moment.modifyHolidays.add({
     "Inauguration Day": {
       date: '1/20',
       keywords_y: ['inauguration']
@@ -139,7 +141,116 @@ test('modifyHolidays_add', function(t){
 });
 
 test('modifyHolidays_remove', function(t){
-  moment().modifyHolidays.remove('Christmas');
+  moment.modifyHolidays.set('US').remove('Christmas');
   var w = moment().holiday('Christmas');
   t.false(w);
+});
+
+test('easter', function(t){
+  moment.modifyHolidays.add('Easter');
+  var w = moment('2016-10-20').holiday('Easter Sunday');
+  t.true(moment.isMoment(w));
+  t.true(w.isSame('2016-03-27', 'day'));
+});
+
+test('Canada', function(t){
+  moment.modifyHolidays.set('Canada').add('Easter');
+  var w = moment('2001-12-26').isHoliday('Boxing Day');
+  t.true(w);
+});
+
+test('parser_1', function(t){
+  moment.modifyHolidays.set({ 'test': { date: '10/5', keywords: ['test'] } });
+  var w = moment().holiday('test');
+  t.true(moment.isMoment(w));
+  t.true(w.isSame(moment().month(9).date(5), 'day'));
+});
+
+test('parser_2', function(t){
+  moment.modifyHolidays.set({ 'test': { date: '8/16/2001', keywords: ['test'] } });
+  var w = moment().holiday('test');
+  t.true(moment.isMoment(w));
+  t.true(w.isSame(moment('2001-08-16'), 'day'));
+});
+
+test('parser_3', function(t){
+  moment.modifyHolidays.set({ 'test': { date: '8/(2,1)', keywords: ['test'] } });
+  var w = moment('2009-10-18').holiday('test');
+  t.true(moment.isMoment(w));
+  t.true(w.isSame(moment('2009-08-04'), 'day'));
+});
+
+test('parser_4', function(t){
+  moment.modifyHolidays.set({ 'test': { date: '2/(5,-2)', keywords: ['test'] } });
+  var w = moment('2002-06-02').holiday('test');
+  t.true(moment.isMoment(w));
+  t.true(w.isSame(moment('2002-02-15'), 'day'));
+});
+
+test('parser_5', function(t){
+  moment.modifyHolidays.set({ 'test': { date: '6/(3)', keywords: ['test'] } });
+  var w = moment('2005-04-12').holiday('test');
+  t.is(w.constructor, Array);
+  t.is(w.length, 5);
+  t.true(w[4].isSame('2005-06-29', 'day'));
+});
+
+test('parser_6', function(t){
+  moment.modifyHolidays.set({ 'test': { date: '(4)/2010', keywords: ['test'] } });
+  var w = moment().holiday('test');
+  t.is(w.constructor, Array);
+  t.is(w.length, 52);
+  t.true(w[21].isSame('2010-06-03', 'day'));
+});
+
+test('parser_7', function(t){
+  moment.modifyHolidays.set({ 'test': { date: '4/(5,2)/2003', keywords: ['test'] } });
+  var w = moment().holiday('test');
+  t.true(moment.isMoment(w));
+  t.true(w.isSame(moment('2003-04-11'), 'day'));
+});
+
+test('parser_8', function(t){
+  moment.modifyHolidays.set({ 'test': { date: '12/(0,-1)/2008', keywords: ['test'] } });
+  var w = moment().holiday('test');
+  t.true(moment.isMoment(w));
+  t.true(w.isSame(moment('2008-12-28'), 'day'));
+});
+
+test('parser_9', function(t){
+  moment.modifyHolidays.set({ 'test': { date: '2/14|3/1', keywords: ['test'] } });
+  var w = moment().holiday('test');
+  t.is(w.constructor, Array);
+  t.is(w.length, 16);
+  t.true(w[9].isSame(moment().month(1).date(23), 'day'));
+});
+
+test('parser_10', function(t){
+  moment.modifyHolidays.set({ 'test': { date: '4/(1,1)|4/(5,-1)', keywords: ['test'] } });
+  var w = moment('2005-11-09').holiday('test');
+  t.is(w.constructor, Array);
+  t.is(w.length, 26);
+  t.true(w[11].isSame(moment('2005-04-15'), 'day'));
+});
+
+test('parser_11', function(t){
+  moment.modifyHolidays.set({ 'test': { date: '10/(3,[11])', keywords: ['test'] } });
+  var w = moment('2008-06-16').holiday('test');
+  t.true(moment.isMoment(w));
+  t.true(w.isSame(moment('2008-10-15'), 'day'));
+});
+
+test('parser_12', function(t){
+  moment.modifyHolidays.set({ 'test': { date: '12/(5,[-23])', keywords: ['test'] } });
+  var w = moment('2014-02-20').holiday('test');
+  t.true(moment.isMoment(w));
+  t.true(w.isSame(moment('2014-12-19'), 'day'));
+});
+
+test('parser_13', function(t){
+  moment.modifyHolidays.set({ 'test': { date: '03/(2,[-16])|3/(5,-1)', keywords: ['test'] } });
+  var w = moment('2012-04-25').holiday('test');
+  t.is(w.constructor, Array);
+  t.is(w.length, 18);
+  t.true(w[7].isSame(moment('2012-03-20'), 'day'));
 });