+++ CHANGELOG.md
@@ -6,13 +6,27 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## [Unreleased]
 
 
+## [1.3.0] - TBA
+### Added
+- Support for holidays that span multiple days.
+- Support for "The [Weekday] on or before/after [Day]".
+- Support for year specific holidays.
+- Locale functionality.
+- Easter and Canadian locales.
+- Documentation on `modifyHolidays.extendParser` function.
+
+### Changed
+- `moment().modifyHolidays` to `moment.modifyHolidays`.
+- Made `moment.modifyHolidays` functions chainable.
+
+
 ## [1.2.0] - 2017-07-17
-## Added
+### Added
 - `modifyHolidays.extendParser` function.
 
 
 ## [1.1.1] - 2017-07-17
-## Changed
+### Changed
 - `previousHoliday` and `nextHoliday` functions will return a single moment object if `count` is set to `1`.
 
 
@@ -33,6 +47,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Initial release.
 
 [Unreleased]: https://github.com/kodie/moment-holiday/compare/v1.0.0...HEAD
+[1.3.0]: https://github.com/kodie/moment-holiday/compare/v1.2.1...v1.3.0
 [1.2.0]: https://github.com/kodie/moment-holiday/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/kodie/moment-holiday/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/kodie/moment-holiday/compare/v1.0.0...v1.1.0